+++ tests/Command/GetLoginCommandTest.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Keboola\DeveloperPortal\Cli\Command\Test;
+namespace Keboola\DeveloperPortal\Cli\Tests\Command;
 
-use Keboola\DeveloperPortal\Cli\Command\GetRepository;
-use PHPUnit\Framework\TestCase;
+use Keboola\DeveloperPortal\Cli\Command\GetLoginCommand;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
+use PHPUnit\Framework\TestCase;
 
-class GetRepositoryTest extends TestCase
+class GetLoginCommandTest extends TestCase
 {
     public function testExecute(): void
     {
         $application = new Application();
-        $application->add(new GetRepository());
-        $command = $application->find('ecr:get-repository');
+        $application->add(new GetLoginCommand());
+        $command = $application->find('ecr:get-login');
         $commandTester = new CommandTester($command);
         $commandTester->execute(array(
             'command'  => $command->getName(),
@@ -23,7 +23,6 @@ class GetRepositoryTest extends TestCase
             'app' => getenv('KBC_DEVELOPERPORTAL_TEST_APP')
         ));
         $this->assertEquals(0, $commandTester->getStatusCode());
-        $this->assertContains(getenv('KBC_DEVELOPERPORTAL_TEST_APP'), $commandTester->getDisplay());
-        $this->assertContains('ecr', $commandTester->getDisplay());
+        $this->assertContains('docker login -u AWS -p', $commandTester->getDisplay());
     }
 }
